[BW]: [Fix] CI scripts check execution status of imporant steps

+++ gitlab-ci/build.sh
@@ -6,8 +6,15 @@ ssh-add <(echo "$SSH_PRIVATE_KEY")
 git submodule update --init --recursive
 cd cd-scripts
 cp config-$CI_BEOS_CONFIG_NAME.py config.py
-./deploy.py --build-beos
+if ! ./deploy.py --build-beos
+then
+  cat beos_deploy_main.log
+  printf "Unable to build beos project. Exiting..."
+  exit 1
+fi
+
 cat beos_deploy_main.log
+
 echo "Removing object files"
 find ~/ci/build/$CI_COMMIT_REF_NAME -type f -name '*.o' -delete
 echo "Stopping old keosd instance"
@@ -16,5 +23,12 @@ python3 ./wait.py $(lsof -t -i:8900)
 echo "Stopping old nodeos instance"
 kill -2 $(lsof -t -i:8888) || true
 python3 ./wait.py $(lsof -t -i:8888)
-./deploy.py --initialize-beos
+
+if ! ./deploy.py --initialize-beos
+  cat beos_deploy_main.log
+  printf "Unable to initialize beos blockchain"
+  exit 2
+fi
+
 cat beos_deploy_main.log
+
